update main bot

+++ uno.py
@@ -5,8 +5,9 @@ import asyncio
 import random
 import traceback
 import json
+from typing import Optional
 
-bot: commands.Bot|None = None # This should be overwritten by the importing script
+bot: commands.Bot # This should be overwritten by the importing script
 cards = ['r0', 'r1', 'r1', 'r2', 'r2', 'r3', 'r3', 'r4', 'r4', 'r5', 'r5', 'r6', 'r6', 'r7', 'r7', 'r8', 'r8', 'r9', 'r9', 'rx', 'rx', 'rr', 'rr', 'r+', 'r+', 'g0', 'g1', 'g1', 'g2', 'g2', 'g3', 'g3', 'g4', 'g4', 'g5', 'g5', 'g6', 'g6', 'g7', 'g7', 'g8', 'g8', 'g9', 'g9', 'gx', 'gx', 'gr', 'gr', 'g+', 'g+', 'b0', 'b1', 'b1', 'b2', 'b2', 'b3', 'b3', 'b4', 'b4', 'b5', 'b5', 'b6', 'b6', 'b7', 'b7', 'b8', 'b8', 'b9', 'b9', 'bx', 'bx', 'br', 'br', 'b+', 'b+', 'y0', 'y1', 'y1', 'y2', 'y2', 'y3', 'y3', 'y4', 'y4', 'y5', 'y5', 'y6', 'y6', 'y7', 'y7', 'y8', 'y8', 'y9', 'y9', 'yx', 'yx', 'yr', 'yr', 'y+', 'y+', 'c?', 'c?', 'c?', 'c?', 'c*', 'c*', 'c*', 'c*']
 
 
@@ -129,13 +130,21 @@ class Player():
             self.game.deck = self.game.stack[:-1]
             self.game.stack = [self.game.stack[-1]]
     
+    async def refreshButtons(self, interaction: discord.Interaction):
+        self.drawCardButton = self.DrawButton(self)
+        self.nextPlayerButton = self.NextButton(self)
+        for card in self.cards:
+            card.refreshButton()
+        await self.send_interaction(False, interaction)
+    
     @handleCrashes("game.message")
-    async def send_interaction(self):
+    async def send_interaction(self, editGlobalMessage: bool = True, interaction: Optional[discord.Interaction] = None):
         if self.game.ended:
             winner = self.game.players[[len(player.cards) == 0 for player in self.game.gamePlayers].index(True)]
             await self.game.close(f"{winner.mention} has won the Game!")
             for player in self.game.gamePlayers:
-                await player.message.delete()
+                if player.message:
+                    await player.message.delete()
             return
         view = discord.ui.View()
         if self.game.draw:
@@ -147,8 +156,16 @@ class Player():
         for card in self.cards:
             view.add_item(card.button)
         sep = "\n"
-        await self.game.message.edit(content=f"{self.game.creator.display_name} has created a game of Uno!\n\nCurrent Card: { {'r': 'Red', 'g': 'Green', 'b': 'Blue', 'y': 'Gray'}[self.game.stack[-1].selectedColor if self.game.stack[-1].selectedColor in ['r', 'g', 'b', 'y'] else self.game.stack[-1].color]} {self.game.stack[-1].symbol}\nIt's {self.game.gamePlayers[self.game.currentPlayer].user.mention}s turn.\n{sep.join([f'<a:alarm:1373945129332768830> {player.user.display_name} has only 1 card left! <a:alarm:1373945129332768830>' for player in self.game.gamePlayers if len(player.cards) == 1])}", view=None)
-        await self.message.edit(content=f"Your Cards:", view=view)
+        if not self.game.message:
+            raise ValueError
+        if editGlobalMessage:
+            globalView = discord.ui.View()
+            globalView.add_item(RefreshCardsButton(self.game))
+            await self.game.message.edit(content=f"{self.game.creator.display_name} has created a game of Uno!\n\nCurrent Card: { {'r': 'Red', 'g': 'Green', 'b': 'Blue', 'y': 'Gray'}[self.game.stack[-1].selectedColor if self.game.stack[-1].selectedColor in ['r', 'g', 'b', 'y'] else self.game.stack[-1].color]} {self.game.stack[-1].symbol}\nIt's {self.game.gamePlayers[self.game.currentPlayer].user.mention}s turn.\n{sep.join([f'<a:alarm:1373945129332768830> {player.user.display_name} has only 1 card left! <a:alarm:1373945129332768830>' for player in self.game.gamePlayers if len(player.cards) == 1])}", view=globalView)
+        if self.message:
+            await self.message.edit(content=f"Your Cards:", view=view)
+        elif interaction:
+            self.message = await interaction.followup.send(content="Your Cards:", view=view, ephemeral=True)
 
 class Card():
     class Button(discord.ui.Button):
@@ -158,6 +175,8 @@ class Card():
         
         @handleCrashes("card.game.message")
         async def callback(self, interaction: discord.Interaction):
+            if not self.card.owner:
+                raise ValueError
             self.card.owner.lastInteraction = interaction
             if not self.card.owner == self.card.game.gamePlayers[self.card.game.currentPlayer]:
                 await interaction.response.send_message("It's not your Turn!", ephemeral=True)
@@ -200,7 +219,8 @@ class Card():
                 for player in self.card.game.gamePlayers:
                     await player.send_interaction()
             else:
-                await self.card.owner.send_interaction()
+                if self.card.owner:
+                    await self.card.owner.send_interaction()
             self.card.owner = None
             
     class Action():
@@ -249,12 +269,17 @@ class Card():
         self.selectedColor: str|None = None
     
     def useable(self) -> bool:
+        if not self.owner:
+            raise ValueError
         lastCard = self.owner.game.stack[-1]
         if self.game.draw != 0:
             return (lastCard.value == "+" and self.value == "+") or (lastCard.value == "+" and self.value == "*" and self.game.settings.setting.plus4onplus2) or (lastCard.value == "*" and self.value == "*" and self.game.settings.setting.plus4onplus4)
         if lastCard.color != "c":
             return self.color == lastCard.color or self.value == lastCard.value or self.color == 'c'
         return self.color == "c" or self.color == lastCard.selectedColor
+    
+    def refreshButton(self):
+        self.button = self.Button(self.buttonStyle, self.symbol, self)
 
 class Settings():
     class setting():
@@ -271,7 +296,8 @@ class Settings():
         async def callback(self, interaction: discord.Interaction):
             await interaction.response.defer()
             setattr(self.settingsmenu.setting, self.setting, not self.settingsmenu.setting.__dict__[self.setting])
-            await self.settingsmenu.message.edit(content=self.settingsmenu.get_message())
+            if self.settingsmenu.message:
+                await self.settingsmenu.message.edit(content=self.settingsmenu.get_message())
 
     def __init__(self, game: "Uno"):
         self.game = game
@@ -335,7 +361,8 @@ class Uno():
                 return
         if self.settings.message:
             await self.settings.message.delete()
-        await self.message.edit(content = f"{self.creator.display_name} has create a game of Uno!\n\nStarting...", view=None)
+        if self.message:
+            await self.message.edit(content = f"{self.creator.display_name} has create a game of Uno!\n\nStarting...", view=None)
         self.running = True
         self.gamePlayers = [Player(player, self) for player in self.players]
         self.currentPlayer = random.randint(0, len(self.gamePlayers)-1)
@@ -357,7 +384,8 @@ class Uno():
 
     @handleCrashes("message")
     async def close(self, message: str):
-        await self.message.edit(content=f"{self.creator.display_name} has created a game of Uno!\n\n{message}", view=None)
+        if self.message:
+            await self.message.edit(content=f"{self.creator.display_name} has created a game of Uno!\n\n{message}", view=None)
         del self
 
     def next_player(self):
@@ -382,7 +410,8 @@ class JoinButton(discord.ui.Button):
         else:
             self.game.addUser(interaction)
             await interaction.response.defer()
-            await self.game.message.edit(content=self.game.lobbyMessage())
+            if self.game.message:
+                await self.game.message.edit(content=self.game.lobbyMessage())
 
 class LeaveButton(discord.ui.Button):
     def __init__(self, game: Uno):
@@ -394,6 +423,8 @@ class LeaveButton(discord.ui.Button):
         if interaction.user not in self.game.players:
             await interaction.response.send_message(f"You are not in this game!", ephemeral=True)
         else:
+            if not self.game.message:
+                raise ValueError
             await interaction.response.defer()
             index = self.game.players.index(interaction.user)
             self.game.players.pop(index)
@@ -414,6 +445,8 @@ class ReadyButton(discord.ui.Button):
 
     @handleCrashes("game.message")
     async def callback(self, interaction: discord.Interaction):
+        if not self.game.message:
+            raise ValueError
         if interaction.user in self.game.players:
             await interaction.response.defer()
             self.game.playerReady[interaction.user.id] = not self.game.playerReady[interaction.user.id]
@@ -443,6 +476,19 @@ class SettingsButton(discord.ui.Button):
         else:
             await interaction.response.send_message(f"You are not in the game!", ephemeral=True)
 
+class RefreshCardsButton(discord.ui.Button):
+    def __init__(self, game: Uno):
+        self.game = game
+        super().__init__(label="Refresh your Message", style=discord.ButtonStyle.gray)
+    
+    @handleCrashes("game.message")
+    async def callback(self, interaction: discord.Interaction):
+        if not interaction.user in self.game.players:
+            await interaction.response.send_message(f"You are not in the game!", ephemeral=True)
+            return
+        await interaction.response.defer()
+        await self.game.gamePlayers[self.game.players.index(interaction.user)].refreshButtons(interaction)
+
 class UnoCommand(discord.app_commands.Group):
     def __init__(self):
         super().__init__(name="uno", description="UNO")
@@ -452,4 +498,5 @@ class UnoCommand(discord.app_commands.Group):
         game = Uno(interaction)
         game.addUser(interaction)
         await interaction.response.send_message("Opening Game...", delete_after=0.0, ephemeral=True)
-        game.message = await interaction.channel.send(game.lobbyMessage(), view=game.view)
+        if interaction.channel and isinstance(interaction.channel, (discord.VoiceChannel, discord.StageChannel, discord.TextChannel, discord.Thread, discord.DMChannel, discord.GroupChannel)):
+            game.message = await interaction.channel.send(game.lobbyMessage(), view=game.view)
